NowPlaying.js: reduce tick() pileup
rewrite the useEffect tick() now playing ':' animation
to only fire once per second. the timeout function shouldn't
be re-scheduling itself, so rely on useEffect to reschedule
the tick.

the previous implementation would pile up multiple calls to tick()
which would trigger a waterfall of re-renders causing CPU usage
to spike and continue to increase the longer the page was
left open.

this change improves battery life on mobile