App.js: always set autoPlay="true"
Always set autoPlay="true" for the audio tag. This autoplay determines
whether the clip starts playing when the source is changed.

I consider it to be independent of the toggle button autoplay, which
means, continue playing calls from the filteredCalls list without
additional input.

With autoPlay="true" the up/down hot keys and tapping calls in the list
immediately starts playback without having to also hit the play
button on the player controls.